requirements.txt
annotated-types==0.7.0
app-model==0.3.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
controller==0.1.0
Flask==3.0.3
Flask-MySQL==1.5.2
Flask-WTF==1.2.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
in-n-out==0.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysql-connector-python-rf==2.2.2
packaging==24.1
psygnal==0.11.1
pydantic==2.9.2
pydantic-compat==0.1.2
pydantic_core==2.23.4
PyMySQL==1.1.1
python-engineio==4.9.1
python-socketio==5.11.4
requests==2.32.3
simple-websocket==1.0.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
WTForms==3.1.2
